+++ lib/websockex.ex
@@ -82,14 +82,6 @@ defmodule WebSockex do
   @typedoc """
   Options values for `start_link`.
 
-  - `:async` - Replies with `{:ok, pid}` before establishing the connection.
-    This is useful for when attempting to connect indefinitely, this way the
-    process doesn't block trying to establish a connection.
-  - `:handle_initial_conn_failure` - When set to `true` a connection failure
-    while establishing the initial connection won't immediately return an error
-    and instead will invoke the `c:handle_disconnect/2` callback. This option
-    only matters during process initialization. The `handle_disconnect`
-    callback is always invoked if an established connection is lost.
   - `:debug` - Options to set the debug options for `:sys.handle_debug`.
   - `:name` - An atom that the registers the process with name locally.
     Can also be a `{:via, module, term}` or `{:global, term}` tuple.
@@ -98,10 +90,8 @@ defmodule WebSockex do
   """
   @type option ::
           WebSockex.Conn.connection_option()
-          | {:async, boolean}
           | {:debug, debug_opts}
           | {:name, atom | {:global, term} | {:via, module, term}}
-          | {:handle_initial_conn_failure, boolean}
 
   @typedoc """
   The reason a connection was closed.
@@ -138,6 +128,11 @@ defmodule WebSockex do
           conn: WebSockex.Conn.t()
         }
 
+  @typedoc """
+  A key used to correlate a result with a send request
+  """
+  @type send_key :: term
+
   @doc """
   Invoked after a connection is established.
 
@@ -154,6 +149,7 @@ defmodule WebSockex do
   @callback handle_frame(frame, state :: term) ::
               {:ok, new_state}
               | {:reply, frame, new_state}
+              | {:reply, frame, send_key, new_state}
               | {:close, new_state}
               | {:close, close_frame, new_state}
             when new_state: term
@@ -164,6 +160,7 @@ defmodule WebSockex do
   @callback handle_cast(msg :: term, state :: term) ::
               {:ok, new_state}
               | {:reply, frame, new_state}
+              | {:reply, frame, send_key, new_state}
               | {:close, new_state}
               | {:close, close_frame, new_state}
             when new_state: term
@@ -174,9 +171,20 @@ defmodule WebSockex do
   @callback handle_info(msg :: term, state :: term) ::
               {:ok, new_state}
               | {:reply, frame, new_state}
+              | {:reply, frame, send_key, new_state}
               | {:close, new_state}
               | {:close, close_frame, new_state}
             when new_state: term
+
+  @doc """
+  Invoked to handle all results for sending messages via `:reply`
+  """
+  @callback handle_send_result(result, frame, send_key, state :: term) ::
+              {:ok, new_state}
+              | {:close, new_state}
+              | {:close, close_frame, new_state}
+            when new_state: term,
+                 result: :ok | {:error, term}
 
   @doc """
   Invoked when the WebSocket disconnects from the server.
@@ -185,24 +193,20 @@ defmodule WebSockex do
   of crashes or other errors the process will terminate immediately
   skipping this callback.
 
-  If the `handle_initial_conn_failure: true` option is provided during process
-  startup, then this callback will be invoked if the process fails to establish
-  an initial connection.
-
   If a connection is established by reconnecting, the `c:handle_connect/2`
   callback will be invoked.
 
   The possible returns for this callback are:
 
   - `{:ok, state}` will continue the process termination.
-  - `{:reconnect, state}` will attempt to reconnect instead of terminating.
-  - `{:reconnect, conn, state}` will attempt to reconnect with the connection
+  - `{:backoff, timeout, state}` will attempt to reconnect after timeout instead of terminating.
+  - `{:backoff, timeout, conn, state}` will attempt to reconnect after timeout with the connection
     data in `conn`. `conn` is expected to be a `t:WebSockex.Conn.t/0`.
   """
   @callback handle_disconnect(connection_status_map, state :: term) ::
               {:ok, new_state}
-              | {:reconnect, new_state}
-              | {:reconnect, new_conn :: WebSockex.Conn.t(), new_state}
+              | {:backoff, timeout(), new_state}
+              | {:backoff, timeout(), new_conn :: WebSockex.Conn.t(), new_state}
             when new_state: term
 
   @doc """
@@ -253,6 +257,7 @@ defmodule WebSockex do
   defmacro __using__(opts) do
     quote location: :keep do
       @behaviour WebSockex
+      require Logger
 
       if Kernel.function_exported?(Supervisor, :child_spec, 2) do
         @doc false
@@ -304,6 +309,17 @@ defmodule WebSockex do
       end
 
       @doc false
+      def handle_send_result(:ok, frame, key, state) do
+        {:ok, state}
+      end
+
+      @doc false
+      def handle_send_result({:error, reason}, frame, key, state) do
+        Logger.warn("[WebSockex] A message failed to send because: #{inspect(reason)}")
+        {:close, state}
+      end
+
+      @doc false
       def handle_ping(:ping, state) do
         {:reply, :pong, state}
       end
@@ -329,6 +345,7 @@ defmodule WebSockex do
                      handle_ping: 2,
                      handle_pong: 2,
                      handle_disconnect: 2,
+                     handle_send_result: 4,
                      terminate: 2,
                      code_change: 3
     end
@@ -452,16 +469,8 @@ defmodule WebSockex do
   ## OTP Stuffs
 
   @doc false
-  def system_continue(parent, debug, %{connection_status: :connected} = state) do
-    websocket_loop(parent, debug, Map.delete(state, :connection_status))
-  end
-
-  def system_continue(parent, debug, %{connection_status: :connecting} = state) do
-    open_loop(parent, debug, Map.delete(state, :connection_status))
-  end
-
-  def system_continue(parent, debug, %{connection_status: {:closing, reason}} = state) do
-    close_loop(reason, parent, debug, Map.delete(state, :connection_status))
+  def system_continue(parent, debug, state) do
+    loop(parent, debug, state)
   end
 
   @doc false
@@ -543,64 +552,33 @@ defmodule WebSockex do
     # OTP stuffs
     debug = Utils.parse_debug_options(self(), opts)
 
-    reply_fun =
-      case Keyword.get(opts, :async, false) do
-        true ->
-          :proc_lib.init_ack(parent, {:ok, self()})
-          &async_init_fun/1
-
-        false ->
-          &sync_init_fun(parent, &1)
-      end
+    :proc_lib.init_ack(parent, {:ok, self()})
 
     state = %{
       conn: conn,
       module: module,
       module_state: module_state,
       name: name,
-      reply_fun: reply_fun,
       buffer: <<>>,
-      fragment: nil
+      fragment: nil,
+      backoff: nil,
+      attempt_number: 0
     }
 
-    handle_conn_failure = Keyword.get(opts, :handle_initial_conn_failure, false)
-
-    case open_connection(parent, debug, state) do
-      {:ok, new_state} ->
-        debug = Utils.sys_debug(debug, :connected, state)
-        module_init(parent, debug, new_state)
-
-      {:error, error, new_state} when handle_conn_failure == true ->
-        init_conn_failure(error, parent, debug, new_state)
-
-      {:error, error, _} ->
-        state.reply_fun.({:error, error})
-    end
+    connect(parent, debug, state)
   end
 
   # Loops
 
-  defp open_loop(parent, debug, state) do
+  defp loop(parent, debug, %{connection_status: :connecting} = state) do
     %{task: %{ref: ref}} = state
 
     receive do
       {:system, from, req} ->
-        state = Map.put(state, :connection_status, :connecting)
         :sys.handle_system_msg(req, from, parent, __MODULE__, debug, state)
 
-      {:"$websockex_send", from, _frame} ->
-        :gen.reply(from, {:error, %WebSockex.NotConnectedError{connection_state: :opening}})
-        open_loop(parent, debug, state)
-
       {:EXIT, ^parent, reason} ->
-        case state do
-          %{reply_fun: reply_fun} ->
-            reply_fun.(reason)
-            exit(reason)
-
-          _ ->
-            terminate(reason, parent, debug, state)
-        end
+        terminate(reason, parent, debug, state)
 
       {^ref, {:ok, new_conn}} ->
         Process.demonitor(ref, [:flush])
@@ -618,7 +596,34 @@ defmodule WebSockex do
     end
   end
 
-  defp websocket_loop(parent, debug, state) do
+  defp loop(parent, debug, %{connection_status: :backoff} = state) do
+    %{backoff: backoff} = state
+
+    receive do
+      {:system, from, req} ->
+        :sys.handle_system_msg(req, from, parent, __MODULE__, debug, state)
+
+      {:"$websockex_send", from, _frame} ->
+        :gen.reply(from, {:error, %WebSockex.NotConnectedError{connection_state: :backoff}})
+        loop(parent, debug, state)
+
+      {:"$websockex_cast", msg} ->
+        debug = Utils.sys_debug(debug, {:in, :cast, msg}, state)
+        common_handle({:handle_cast, msg}, parent, debug, state)
+
+      {:EXIT, ^parent, reason} ->
+        terminate(reason, parent, debug, state)
+
+      {:timeout, ^backoff, :backoff} ->
+        connect(parent, debug, %{state | backoff: nil})
+
+      msg ->
+        debug = Utils.sys_debug(debug, {:in, :msg, msg}, state)
+        common_handle({:handle_info, msg}, parent, debug, state)
+    end
+  end
+
+  defp loop(parent, debug, %{connection_status: :connected} = state) do
     case WebSockex.Frame.parse_frame(state.buffer) do
       {:ok, frame, buffer} ->
         debug = Utils.sys_debug(debug, {:in, :frame, frame}, state)
@@ -633,7 +638,6 @@ defmodule WebSockex do
 
         receive do
           {:system, from, req} ->
-            state = Map.put(state, :connection_status, :connected)
             :sys.handle_system_msg(req, from, parent, __MODULE__, debug, state)
 
           {:"$websockex_cast", msg} ->
@@ -645,7 +649,7 @@ defmodule WebSockex do
 
           {^transport, ^socket, message} ->
             buffer = <<state.buffer::bitstring, message::bitstring>>
-            websocket_loop(parent, debug, %{state | buffer: buffer})
+            loop(parent, debug, %{state | buffer: buffer})
 
           {:tcp_closed, ^socket} ->
             handle_close({:remote, :closed}, parent, debug, state)
@@ -663,9 +667,14 @@ defmodule WebSockex do
     end
   end
 
-  defp close_loop(reason, parent, debug, %{conn: conn, timer_ref: timer_ref} = state) do
+  defp loop(
+         parent,
+         debug,
+         %{conn: conn, timer_ref: timer_ref, connection_status: {:closing, reason}} = state
+       ) do
     transport = state.conn.transport
     socket = state.conn.socket
+    state = cancel_backoff(state)
 
     receive do
       {:system, from, req} ->
@@ -676,11 +685,11 @@ defmodule WebSockex do
         terminate(reason, parent, debug, state)
 
       {^transport, ^socket, _} ->
-        close_loop(reason, parent, debug, state)
+        loop(parent, debug, state)
 
       {:"$websockex_send", from, _frame} ->
         :gen.reply(from, {:error, %WebSockex.NotConnectedError{connection_state: :closing}})
-        close_loop(reason, parent, debug, state)
+        loop(parent, debug, state)
 
       {close_mod, ^socket} when close_mod in [:tcp_closed, :ssl_closed] ->
         new_conn = %{conn | socket: nil}
@@ -739,7 +748,7 @@ defmodule WebSockex do
   end
 
   defp handle_fragment({:fragment, type, part}, parent, debug, %{fragment: nil} = state) do
-    websocket_loop(parent, debug, %{state | fragment: {type, part}})
+    loop(parent, debug, %{state | fragment: {type, part}})
   end
 
   defp handle_fragment({:fragment, _, _}, parent, debug, state) do
@@ -761,7 +770,7 @@ defmodule WebSockex do
   end
 
   defp handle_fragment({:continuation, next}, parent, debug, %{fragment: {type, part}} = state) do
-    websocket_loop(parent, debug, %{state | fragment: {type, <<part::binary, next::binary>>}})
+    loop(parent, debug, %{state | fragment: {type, <<part::binary, next::binary>>}})
   end
 
   defp handle_fragment({:finish, next}, parent, debug, %{fragment: {type, part}} = state) do
@@ -801,22 +810,15 @@ defmodule WebSockex do
 
     case result do
       {:ok, new_state} ->
-        websocket_loop(parent, debug, %{state | module_state: new_state})
+        loop(parent, debug, %{state | module_state: new_state})
 
       {:reply, frame, new_state} ->
-        # A `with` that includes `else` clause isn't tail recursive (elixir-lang/elixir#6251)
-        res =
-          with {:ok, binary_frame} <- WebSockex.Frame.encode_frame(frame),
-               do: WebSockex.Conn.socket_send(state.conn, binary_frame)
+        debug = Utils.sys_debug(debug, {:reply, function, frame, nil}, state)
+        do_send(frame, nil, parent, debug, %{state | module_state: new_state})
 
-        case res do
-          :ok ->
-            debug = Utils.sys_debug(debug, {:reply, function, frame}, state)
-            websocket_loop(parent, debug, %{state | module_state: new_state})
-
-          {:error, error} ->
-            handle_close({:error, error}, parent, debug, %{state | module_state: new_state})
-        end
+      {:reply, frame, ref, new_state} ->
+        debug = Utils.sys_debug(debug, {:reply, function, frame, ref}, state)
+        do_send(frame, ref, parent, debug, %{state | module_state: new_state})
 
       {:close, new_state} ->
         handle_close({:local, :normal}, parent, debug, %{state | module_state: new_state})
@@ -853,7 +855,12 @@ defmodule WebSockex do
       end
 
     timer_ref = Process.send_after(self(), :"$websockex_close_timeout", 5000)
-    close_loop(reason, parent, debug, Map.put(state, :timer_ref, timer_ref))
+
+    loop(
+      parent,
+      debug,
+      Map.merge(state, %{timer_ref: timer_ref, connection_status: {:closing, reason}})
+    )
   end
 
   defp handle_local_close(reason, parent, debug, state) do
@@ -863,7 +870,12 @@ defmodule WebSockex do
       :ok ->
         debug = Utils.sys_debug(debug, {:socket_out, :close, reason}, state)
         timer_ref = Process.send_after(self(), :"$websockex_close_timeout", 5000)
-        close_loop(reason, parent, debug, Map.put(state, :timer_ref, timer_ref))
+
+        loop(
+          parent,
+          debug,
+          Map.merge(state, %{timer_ref: timer_ref, connection_status: {:closing, reason}})
+        )
 
       {:error, %WebSockex.ConnError{original: reason}} when reason in [:closed, :einval] ->
         handle_close({:remote, :closed}, parent, debug, state)
@@ -874,7 +886,12 @@ defmodule WebSockex do
     send_close_frame(:error, state.conn)
 
     timer_ref = Process.send_after(self(), :"$websockex_close_timeout", 5000)
-    close_loop(reason, parent, debug, Map.put(state, :timer_ref, timer_ref))
+
+    loop(
+      parent,
+      debug,
+      Map.merge(state, %{timer_ref: timer_ref, connection_status: {:closing, reason}})
+    )
   end
 
   @spec handle_terminate_close(any, pid, any, any) :: no_return
@@ -906,7 +923,7 @@ defmodule WebSockex do
 
         :gen.reply(from, :ok)
         debug = Utils.sys_debug(debug, {:socket_out, :sync_send, frame}, state)
-        websocket_loop(parent, debug, state)
+        loop(parent, debug, state)
 
       {:error, %WebSockex.ConnError{original: reason}} = error
       when reason in [:closed, :einval] ->
@@ -915,7 +932,34 @@ defmodule WebSockex do
 
       {:error, _} = error ->
         :gen.reply(from, error)
-        websocket_loop(parent, debug, state)
+        loop(parent, debug, state)
+    end
+  end
+
+  defp do_send(frame, key, parent, debug, %{connection_status: connection_status} = state) do
+    # A `with` that includes `else` clause isn't tail recursive (elixir-lang/elixir#6251)
+    res =
+      with true <-
+             connection_status == :connected ||
+               {:error, %WebSockex.NotConnectedError{connection_state: connection_status}},
+           {:ok, binary_frame} <- WebSockex.Frame.encode_frame(frame),
+           do: WebSockex.Conn.socket_send(state.conn, binary_frame)
+
+    case handle_send_result(res, frame, key, state) do
+      {:ok, new_state} ->
+        loop(parent, debug, %{state | module_state: new_state})
+
+      {:close, new_state} ->
+        handle_close({:local, :normal}, parent, debug, %{state | module_state: new_state})
+
+      {:close, {close_code, message}, new_state} ->
+        handle_close({:local, close_code, message}, parent, debug, %{
+          state
+          | module_state: new_state
+        })
+
+      {:"$EXIT", reason} ->
+        terminate(reason, parent, debug, state)
     end
   end
 
@@ -938,62 +982,58 @@ defmodule WebSockex do
 
   # Connection Handling
 
-  defp init_conn_failure(reason, parent, debug, state, attempt \\ 1) do
-    case handle_disconnect(reason, state, attempt) do
-      {:ok, new_module_state} ->
-        init_failure(reason, parent, debug, %{state | module_state: new_module_state})
-
-      {:reconnect, new_conn, new_module_state} ->
-        state = %{state | conn: new_conn, module_state: new_module_state}
-        debug = Utils.sys_debug(debug, :reconnect, state)
-
-        case open_connection(parent, debug, state) do
-          {:ok, new_state} ->
-            debug = Utils.sys_debug(debug, :connected, state)
-            module_init(parent, debug, new_state)
-
-          {:error, new_reason, new_state} ->
-            init_conn_failure(new_reason, parent, debug, new_state, attempt + 1)
-        end
-
-      {:"$EXIT", reason} ->
-        init_failure(reason, parent, debug, state)
-    end
-  end
-
-  defp on_disconnect(reason, parent, debug, state, attempt \\ 1) do
-    case handle_disconnect(reason, state, attempt) do
+  defp on_disconnect(reason, parent, debug, state) do
+    case handle_disconnect(reason, state) do
       {:ok, new_module_state} when is_tuple(reason) and elem(reason, 0) == :error ->
         terminate(elem(reason, 1), parent, debug, %{state | module_state: new_module_state})
 
       {:ok, new_module_state} ->
         terminate(reason, parent, debug, %{state | module_state: new_module_state})
 
-      {:reconnect, new_conn, new_module_state} ->
-        state = %{state | conn: new_conn, module_state: new_module_state}
-        debug = Utils.sys_debug(debug, :reconnect, state)
+      {:backoff, backoff_timeout, new_conn, new_module_state} ->
+        backoff = start_backoff(backoff_timeout)
 
-        case open_connection(parent, debug, state) do
-          {:ok, new_state} ->
-            debug = Utils.sys_debug(debug, :reconnected, state)
-            reconnect(parent, debug, new_state)
+        state = %{
+          state
+          | conn: new_conn,
+            backoff: backoff,
+            module_state: new_module_state,
+            connection_status: :backoff
+        }
 
-          {:error, new_reason, new_state} ->
-            on_disconnect(new_reason, parent, debug, new_state, attempt + 1)
-        end
+        loop(parent, debug, state)
 
       {:"$EXIT", reason} ->
         terminate(reason, parent, debug, state)
     end
   end
 
-  defp reconnect(parent, debug, state) do
+  defp connect(parent, debug, state) do
+    case open_connection(parent, debug, state) do
+      {:ok, new_state} ->
+        debug = Utils.sys_debug(debug, :connected, state)
+        on_connected(parent, debug, new_state)
+
+      {:error, error, new_state} ->
+        on_disconnect(error, parent, debug, update_in(new_state.attempt_number, &(&1 + 1)))
+    end
+  end
+
+  defp on_connected(parent, debug, state) do
     result = try_callback(state.module, :handle_connect, [state.conn, state.module_state])
 
     case result do
       {:ok, new_module_state} ->
-        state = Map.merge(state, %{buffer: <<>>, fragment: nil, module_state: new_module_state})
-        websocket_loop(parent, debug, state)
+        state =
+          Map.merge(state, %{
+            buffer: <<>>,
+            fragment: nil,
+            module_state: new_module_state,
+            attempt_number: 1
+          })
+
+        execute_telemetry([:websockex, :connected], state)
+        loop(parent, debug, %{state | connection_status: :connected})
 
       {:"$EXIT", reason} ->
         terminate(reason, parent, debug, state)
@@ -1028,43 +1068,10 @@ defmodule WebSockex do
         end
       end)
 
-    open_loop(parent, debug, Map.put(state, :task, task))
+    loop(parent, debug, Map.merge(state, %{task: task, connection_status: :connecting}))
   end
 
   # Other State Functions
-
-  defp module_init(parent, debug, state) do
-    execute_telemetry([:websockex, :connected], state)
-
-    result = try_callback(state.module, :handle_connect, [state.conn, state.module_state])
-
-    case result do
-      {:ok, new_module_state} ->
-        state.reply_fun.({:ok, self()})
-
-        state =
-          Map.put(state, :module_state, new_module_state)
-          |> Map.delete(:reply_fun)
-
-        websocket_loop(parent, debug, state)
-
-      {:"$EXIT", reason} ->
-        state.reply_fun.(reason)
-
-      badreply ->
-        reason =
-          {:error,
-           %WebSockex.BadResponseError{
-             module: state.module,
-             function: :handle_connect,
-             args: [state.conn, state.module_state],
-             response: badreply
-           }}
-
-        state.reply_fun.(reason)
-    end
-  end
-
   @spec terminate(any, pid, any, any) :: no_return
   defp terminate(reason, parent, debug, state) do
     execute_telemetry([:websockex, :terminate], state, %{reason: reason})
@@ -1092,9 +1099,8 @@ defmodule WebSockex do
     end
   end
 
-  defp handle_disconnect(reason, state, attempt) do
-    status_map = %{conn: state.conn, reason: reason, attempt_number: attempt}
-
+  defp handle_disconnect(reason, state) do
+    status_map = %{conn: state.conn, reason: reason, attempt_number: state.attempt_number}
     execute_telemetry([:websockex, :disconnected], state, status_map)
 
     result = try_callback(state.module, :handle_disconnect, [status_map, state.module_state])
@@ -1103,11 +1109,11 @@ defmodule WebSockex do
       {:ok, new_state} ->
         {:ok, new_state}
 
-      {:reconnect, new_state} ->
-        {:reconnect, state.conn, new_state}
+      {:backoff, timeout, new_conn, new_state} ->
+        {:backoff, timeout, new_conn, new_state}
 
-      {:reconnect, new_conn, new_state} ->
-        {:reconnect, new_conn, new_state}
+      {:backoff, timeout, new_state} ->
+        {:backoff, timeout, state.conn, new_state}
 
       {:"$EXIT", _} = res ->
         res
@@ -1118,6 +1124,33 @@ defmodule WebSockex do
            module: state.module,
            function: :handle_disconnect,
            args: [status_map, state.module_state],
+           response: badreply
+         }}
+    end
+  end
+
+  defp handle_send_result(result, frame, key, state) do
+    result = try_callback(state.module, :handle_send_result, [result, frame, key, state.module_state])
+
+    case result do
+      {:ok, new_state} ->
+        {:ok, new_state}
+
+      {:close, new_state} ->
+        {:close, new_state}
+
+      {:close, close_frame, new_state} ->
+        {:close, close_frame, new_state}
+
+      {:"$EXIT", _} = res ->
+        res
+
+      badreply ->
+        {:"$EXIT",
+         %WebSockex.BadResponseError{
+           module: state.module,
+           function: :handle_send_result,
+           args: [result, frame, state.module_state],
            response: badreply
          }}
     end
@@ -1135,21 +1168,6 @@ defmodule WebSockex do
 
     :exit, payload ->
       {:"$EXIT", payload}
-  end
-
-  defp init_failure(reason, _parent, _debug, state) do
-    state.reply_fun.({:error, reason})
-  end
-
-  defp async_init_fun({:ok, _}), do: :noop
-  defp async_init_fun(exit_reason), do: exit(exit_reason)
-
-  defp sync_init_fun(parent, {error, stacktrace}) when is_list(stacktrace) do
-    :proc_lib.init_ack(parent, {:error, error})
-  end
-
-  defp sync_init_fun(parent, reply) do
-    :proc_lib.init_ack(parent, reply)
   end
 
   defp validate_handshake(headers, key) do
@@ -1185,5 +1203,37 @@ defmodule WebSockex do
     end
   else
     defp execute_telemetry(_, _, _ \\ %{}), do: :ok
+  end
+
+  ## backoff helpers
+
+  defp start_backoff(:infinity), do: nil
+
+  defp start_backoff(timeout) do
+    :erlang.start_timer(timeout, self(), :backoff)
+  end
+
+  defp cancel_backoff(%{backoff: nil} = state), do: state
+
+  defp cancel_backoff(%{backoff: backoff} = state) do
+    case :erlang.cancel_timer(backoff) do
+      false ->
+        flush_backoff(backoff)
+
+      _ ->
+        :ok
+    end
+
+    %{state | backoff: nil}
+  end
+
+  defp flush_backoff(backoff) do
+    receive do
+      {:timeout, ^backoff, :backoff} ->
+        :ok
+    after
+      0 ->
+        :ok
+    end
   end
 end